+++ packages/eslint-config/typescript/strict.js
@@ -1,0 +1,6 @@
+module.exports = {
+  extends: [
+    './index',
+    '../rules/typescript-strict',
+  ].map(require.resolve),
+};