+++ packages/spec/jest.config.js
@@ -5,6 +5,6 @@ module.exports = {
   testPathIgnorePatterns: [
     '<rootDir>/dist/',
     '<rootDir>/node_modules/',
-    '<rootDir>/__testFixtures__/',
+    '<rootDir>/__tests__/__fixtures__/',
   ],
 };