+++ packages/spec/CHANGELOG.md
@@ -1,5 +1,10 @@
 # @applint/spec
 
+## 1.2.0
+
+- feat: support strict rules(`react-ts-strict`, `rax-ts-strict`, `common-ts-strict`)
+- chore: add `"postcss-scss` and `"postcss-less` dependencies
+
 ## 1.1.1
 
 - fix: missing package `eslint-plugin-rax-compile-time-miniapp`